- move pages to compenent dir - move router to project root - ad alias to storybook
