+++ components/pages/index.vue
@@ -19,7 +19,6 @@ export default {
     }
   },
   methods: { },
-  computed: { },
   components: {
     LandingLayout,
   },
