+++ config.json
@@ -1,6 +1,7 @@
 {
     "siteTitle": "Frontend Starter Kit",
     "path": {
-        
+        "components": "components/",
+        "page": "components/pages"
     }
 }
