+++ router.js
@@ -1,11 +1,12 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import store from './../store'
+import store from './store'
+import CONFIG from './config.json'
 
 Vue.use(Router)
 
 function view(name) {
-  return resolve => require([`./../pages/${name}.vue`], resolve)
+  return resolve => require([`./${CONFIG.path.page}/${name}.vue`], resolve)
 }
 
 let routes = [
